refactor(foundation): unify ProtocolContract trait + types (A18)
Move the protocol-contract abstraction to
`mockforge-foundation::protocol_contract_types` so consumers can
reference and implement it without depending on deprecated core
modules.

**Moved to foundation:**
- Trait: `ProtocolContract` (async trait with diff/validate/to_json etc.)
- Data types: `ContractOperation`, `OperationType`, `ContractRequest`,
  `ValidationResult`, `ValidationError`, `ContractMetadata`,
  `ChangeClassification`, `ContractError`
- Helpers: `classify_change`, `extract_breaking_changes`

**Stays in core:**
- `ProtocolContractRegistry` (holds `Box<dyn ProtocolContract>`; it's a
  runtime registry, not a data type)
- Per-protocol impls (`GrpcContract`, `KafkaContract`,
  `WebSocketContract`, `MqttContract`) — they use protocol-specific
  libraries
- `compare_contracts` helper, `generate_grpc_drift_report` helper

Core's `contract_drift::protocol_contracts` module re-exports from
foundation. mockforge-http's 4 `mockforge_core::protocol_abstraction::Protocol`
import sites migrated to `mockforge_foundation::protocol::Protocol`.

Drops another 53 deprecation warnings from http (351 → 298). All 574
http + 1818 core + 42 foundation tests pass; workspace clippy clean.
