+++ crates/mockforge-core/src/contract_drift/protocol_contracts.rs
@@ -3,201 +3,19 @@
 //! This module provides a unified interface for contract definitions across different
 //! protocols (HTTP/OpenAPI, gRPC, WebSocket, MQTT, Kafka), enabling consistent drift
 //! detection and analysis regardless of the transport layer.
+//!
+//! The trait and data types are re-exported from
+//! `mockforge-foundation::protocol_contract_types` so consumers can implement
+//! and use them without depending on deprecated core modules.
 
-use crate::ai_contract_diff::{ContractDiffResult, Mismatch, MismatchSeverity};
-use crate::protocol_abstraction::Protocol;
-use serde::{Deserialize, Serialize};
+use crate::ai_contract_diff::{ContractDiffResult, Mismatch};
+use mockforge_foundation::protocol::Protocol;
+pub use mockforge_foundation::protocol_contract_types::{
+    classify_change, extract_breaking_changes, ChangeClassification, ContractError,
+    ContractMetadata, ContractOperation, ContractRequest, OperationType, ProtocolContract,
+    ValidationError, ValidationResult,
+};
 use std::collections::HashMap;
-
-/// Protocol-agnostic contract definition
-///
-/// This trait allows different protocol implementations (HTTP, gRPC, WebSocket, etc.)
-/// to provide a unified interface for contract drift detection.
-#[async_trait::async_trait]
-pub trait ProtocolContract: Send + Sync {
-    /// Get the protocol type this contract represents
-    fn protocol(&self) -> Protocol;
-
-    /// Get a unique identifier for this contract
-    fn contract_id(&self) -> &str;
-
-    /// Get the contract version
-    fn version(&self) -> &str;
-
-    /// Get all operations/methods/topics defined in this contract
-    fn operations(&self) -> Vec<ContractOperation>;
-
-    /// Get a specific operation by identifier
-    fn get_operation(&self, operation_id: &str) -> Option<&ContractOperation>;
-
-    /// Compare this contract with another contract of the same protocol
-    ///
-    /// Returns a `ContractDiffResult` describing the differences
-    async fn diff(&self, other: &dyn ProtocolContract)
-        -> Result<ContractDiffResult, ContractError>;
-
-    /// Validate a request/message against this contract
-    ///
-    /// Returns validation errors if the request doesn't match the contract
-    async fn validate(
-        &self,
-        operation_id: &str,
-        request: &ContractRequest,
-    ) -> Result<ValidationResult, ContractError>;
-
-    /// Get schema information for an operation
-    ///
-    /// Returns a JSON-serializable representation of the schema
-    fn get_schema(&self, operation_id: &str) -> Option<serde_json::Value>;
-
-    /// Serialize the contract to a JSON representation
-    fn to_json(&self) -> Result<serde_json::Value, ContractError>;
-}
-
-/// An operation defined in a contract (method, endpoint, topic, etc.)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ContractOperation {
-    /// Unique identifier for this operation
-    pub id: String,
-    /// Human-readable name
-    pub name: String,
-    /// Operation type (varies by protocol)
-    pub operation_type: OperationType,
-    /// Input schema (request/message schema)
-    pub input_schema: Option<serde_json::Value>,
-    /// Output schema (response/message schema)
-    pub output_schema: Option<serde_json::Value>,
-    /// Metadata (tags, descriptions, etc.)
-    pub metadata: HashMap<String, String>,
-}
-
-/// Type of operation (protocol-specific)
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum OperationType {
-    /// HTTP endpoint (GET, POST, etc.)
-    HttpEndpoint {
-        /// HTTP method (GET, POST, PUT, DELETE, etc.)
-        method: String,
-        /// Endpoint path (e.g., "/api/users")
-        path: String,
-    },
-    /// gRPC method (service.method)
-    GrpcMethod {
-        /// gRPC service name
-        service: String,
-        /// gRPC method name
-        method: String,
-    },
-    /// WebSocket message type
-    WebSocketMessage {
-        /// Message type identifier
-        message_type: String,
-        /// Optional topic or channel name
-        topic: Option<String>,
-    },
-    /// MQTT topic
-    MqttTopic {
-        /// MQTT topic name
-        topic: String,
-        /// Quality of Service level (0, 1, or 2)
-        qos: Option<u8>,
-    },
-    /// Kafka topic
-    KafkaTopic {
-        /// Kafka topic name
-        topic: String,
-        /// Optional key schema identifier (Avro schema ID, etc.)
-        key_schema: Option<String>,
-        /// Optional value schema identifier (Avro schema ID, etc.)
-        value_schema: Option<String>,
-    },
-}
-
-/// Protocol-agnostic request representation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ContractRequest {
-    /// Protocol type
-    pub protocol: Protocol,
-    /// Operation identifier
-    pub operation_id: String,
-    /// Request payload (serialized)
-    pub payload: Vec<u8>,
-    /// Content type or encoding
-    pub content_type: Option<String>,
-    /// Additional metadata (headers, properties, etc.)
-    pub metadata: HashMap<String, String>,
-}
-
-/// Validation result for contract validation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ValidationResult {
-    /// Whether the request is valid
-    pub valid: bool,
-    /// Validation errors (if any)
-    pub errors: Vec<ValidationError>,
-    /// Warnings (non-fatal issues)
-    pub warnings: Vec<String>,
-}
-
-/// Validation error details
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ValidationError {
-    /// Error message
-    pub message: String,
-    /// Path to the field causing the error (JSONPath, field name, etc.)
-    pub path: Option<String>,
-    /// Error code or type
-    pub code: Option<String>,
-}
-
-/// Contract-related errors
-#[derive(Debug, thiserror::Error)]
-pub enum ContractError {
-    /// Contract was not found in the registry
-    #[error("Contract not found: {0}")]
-    NotFound(String),
-    /// Contract format is invalid or cannot be parsed
-    #[error("Invalid contract format: {0}")]
-    InvalidFormat(String),
-    /// Protocol is not supported for contract operations
-    #[error("Unsupported protocol: {0:?}")]
-    UnsupportedProtocol(Protocol),
-    /// Operation was not found in the contract
-    #[error("Operation not found: {0}")]
-    OperationNotFound(String),
-    /// Schema validation failed
-    #[error("Schema validation error: {0}")]
-    SchemaValidation(String),
-    /// JSON serialization/deserialization error
-    #[error("Serialization error: {0}")]
-    Serialization(#[from] serde_json::Error),
-    /// I/O error (file reading, etc.)
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-    /// Other error occurred
-    #[error("Other error: {0}")]
-    Other(String),
-}
-
-/// Contract metadata
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ContractMetadata {
-    /// Contract name
-    pub name: String,
-    /// Contract version
-    pub version: String,
-    /// Protocol type
-    pub protocol: Protocol,
-    /// Description
-    pub description: Option<String>,
-    /// Tags or categories
-    pub tags: Vec<String>,
-    /// Creation timestamp
-    pub created_at: Option<i64>,
-    /// Last update timestamp
-    pub updated_at: Option<i64>,
-}
 
 /// Registry for managing protocol contracts
 pub struct ProtocolContractRegistry {
@@ -265,72 +83,6 @@ pub async fn compare_contracts(
 
     // Use the contract's diff method
     old_contract.diff(new_contract).await
-}
-
-/// Helper function to extract breaking changes from a diff result
-///
-/// Breaking changes are defined as mismatches with Critical or High severity
-/// that indicate structural incompatibilities (missing required fields, type mismatches, etc.)
-pub fn extract_breaking_changes(diff: &ContractDiffResult) -> Vec<&Mismatch> {
-    diff.mismatches
-        .iter()
-        .filter(|m| {
-            matches!(m.severity, MismatchSeverity::Critical | MismatchSeverity::High)
-                && matches!(
-                    m.mismatch_type,
-                    crate::ai_contract_diff::MismatchType::MissingRequiredField
-                        | crate::ai_contract_diff::MismatchType::TypeMismatch
-                        | crate::ai_contract_diff::MismatchType::EndpointNotFound
-                        | crate::ai_contract_diff::MismatchType::MethodNotAllowed
-                        | crate::ai_contract_diff::MismatchType::SchemaMismatch
-                )
-        })
-        .collect()
-}
-
-/// Classification result for a change
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChangeClassification {
-    /// Whether this change is additive (new methods, fields, services)
-    pub is_additive: bool,
-    /// Whether this change is breaking (removed methods, type changes, etc.)
-    pub is_breaking: bool,
-    /// Category of change (e.g., "method_added", "method_removed", "type_changed")
-    pub change_category: Option<String>,
-}
-
-/// Extract change classification from a mismatch
-///
-/// Uses the context field to determine if a change is additive, breaking, or both
-pub fn classify_change(mismatch: &Mismatch) -> ChangeClassification {
-    // Check if classification is already in context (from gRPC diff)
-    let is_additive =
-        mismatch.context.get("is_additive").and_then(|v| v.as_bool()).unwrap_or(false);
-
-    let is_breaking = mismatch.context.get("is_breaking").and_then(|v| v.as_bool()).unwrap_or({
-        // Fallback: infer from severity and type
-        matches!(mismatch.severity, MismatchSeverity::Critical | MismatchSeverity::High)
-            && matches!(
-                mismatch.mismatch_type,
-                crate::ai_contract_diff::MismatchType::MissingRequiredField
-                    | crate::ai_contract_diff::MismatchType::TypeMismatch
-                    | crate::ai_contract_diff::MismatchType::EndpointNotFound
-                    | crate::ai_contract_diff::MismatchType::MethodNotAllowed
-                    | crate::ai_contract_diff::MismatchType::SchemaMismatch
-            )
-    });
-
-    let change_category = mismatch
-        .context
-        .get("change_category")
-        .and_then(|v| v.as_str())
-        .map(|s| s.to_string());
-
-    ChangeClassification {
-        is_additive,
-        is_breaking,
-        change_category,
-    }
 }
 
 /// Generate a per-service+method drift report for gRPC contracts
