+++ crates/mockforge-foundation/src/lib.rs
@@ -18,6 +18,7 @@ pub mod incidents_types;
 pub mod intelligent_behavior;
 pub mod multi_tenant_types;
 pub mod protocol;
+pub mod protocol_contract_types;
 pub mod state_machine;
 pub mod threat_modeling_types;
 pub mod workspace_promotion;
