+++ crates/mockforge-http/src/middleware/drift_tracking.rs
@@ -189,7 +189,7 @@ pub async fn drift_tracking_middleware_with_extensions(
                             after_sample,
                             Some(drift_result.fitness_test_results.clone()), // fitness_test_results
                             drift_result.consumer_impact.clone(),            // affected_consumers
-                            Some(mockforge_core::protocol_abstraction::Protocol::Http), // protocol
+                            Some(mockforge_foundation::protocol::Protocol::Http), // protocol
                         )
                         .await;
